Rewrite ckeditor5-word-count to TypeScript.
