+++ packages/ckeditor5-word-count/package.json
@@ -10,7 +10,7 @@
     "ckeditor5-plugin",
     "ckeditor5-dll"
   ],
-  "main": "src/index.js",
+  "main": "src/index.ts",
   "dependencies": {
     "lodash-es": "^4.17.15",
     "ckeditor5": "^35.3.2"
@@ -30,6 +30,7 @@
     "@ckeditor/ckeditor5-table": "^35.3.2",
     "@ckeditor/ckeditor5-theme-lark": "^35.3.2",
     "@ckeditor/ckeditor5-utils": "^35.3.2",
+    "typescript": "^4.8.4",
     "webpack": "^5.58.1",
     "webpack-cli": "^4.9.0"
   },
@@ -48,13 +49,16 @@
   },
   "files": [
     "lang",
-    "src",
+    "src/**/*.js",
+    "src/**/*.d.ts",
     "theme",
     "build",
     "ckeditor5-metadata.json",
     "CHANGELOG.md"
   ],
   "scripts": {
-    "dll:build": "webpack"
+    "dll:build": "webpack",
+    "build": "tsc -p ./tsconfig.release.json",
+    "postversion": "npm run build"
   }
 }
