+++ packages/ckeditor5-word-count/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "extends": "ckeditor5/tsconfig.json",
+  "include": [
+    "src",
+    "../../typings"
+  ]
+}
