+++ packages/ckeditor5-word-count/tsconfig.release.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../tsconfig.release.json",
+  "include": [
+    "./src/",
+    "../../typings/"
+  ],
+  "exclude": [
+    "./tests/"
+  ]
+}
